Splitting the get_waveform_bytes() logic into two functions
It simply got too messy when everything was in a
single function. The cases are quite different.

Whether you are reading the internal memory or the
screen memory makes quite a difference in how to handle it.

Especially the screen memory can be nasty when you
stop the scpe and scroll outside the end or start of
the waveform. You will not get all 1200 samples back
in that case...